UiBaru/requirements.txt
customtkinter==5.2.0
pillow==10.0.1
packaging==23.1
